Cut SVGs for better animations (Experiment)
Made some cuts on SVGs in Adobe Illustrator. This should look better on animations. What I did specifically is I made some cuts wherever holes are on SVGs. When animating SVGs, if there are holes, they become filled at the beginning of animation, and pop back in at the end. I'm trying to solve that problem.